Moved folder listner to worker service infrasctructure
