--- src/Core/Octopus.Conductor.Application/Interfaces/IMessageSender.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Octopus.Conductor.Application.Interfaces
-{
-    public interface IMessageSender
-    {
-        Task SendMessage(object message);
-    }
-}
